Avoid blocking the main thread while making the CA
Break up the calculations into smaller pieces with `setTimeout`. This
makes the process overall take a little longer, but the user doesn't
experience any blocking: they can always edit the input.

Also, with this change we don't need to debounce the calculation.
Instead, we adjust the drawing's opacity during creation so it seems to
"fade in," resulting in a less jarring experience.