Calculate the generations on a buffer of zeros
Before, the visible limit was the emulator's limit as well, where it
assumes cells outside the limit are always in zero-state. Now, the
visible limit is zero-buffered in the emulator. The two are similar, but
now effects can travel outside the visible limit without causing
reflection effects.

The buffer's width is set such that spurious signals from the
edge-condition cannot travel to the visible portion of the emulation
within the set number of generations.