Make "favoriting" accurate right at page load
Before, we were lazily loading the Favorites store only when the
favorites tab was viewed. This made it so a new user could load the
page, change the language, and load a fresh favorites tab with the
defaults in the correct language.

However, it also meant that "favoriting" an automaton didn't work until
the favorites tab was opened, even for returning users (ie. refreshing
the page forced you to view the favorites tab before letting you
favorite new automata). Also, the heart didn't accurately reflect if the
current automaton was favorited until after the favorites tab was
opened.

Rework things so that the Favorites store loads when the page loads. Now
the hart status is always accurate, and users can favorite automata
before viewing the favorites tab. Also, as a nice side-effect, until a
new user changes the current set of favorites (either by adding their
own, deleting a default, or changing a default's name/description), the
language of the favorites updates automatically when the user changes
the language.