Fix Hydration error because of currentLocalePath
`currentLocalePath` failed during site generation because there is no
`window` object, so `CurrentPageInLocale` rendered an empty fragment.
Once the app rendered in the browser, though, `currentLocalePath`
worked, and `CurrentPageInLocale` rendered a link. This generated the
error:

    Warning: Expected server HTML to contain a matching <a> in <p>.

Instead of using `currentLocalePath`, use the React hook that nextJS
provides for this situation, `useRouter`.